User Profile Settings moved to UserHome.tsx. CSS added.
