--- apps/app/src/pages/usersetting.tsx
@@ -1,3 +1,0 @@
-import UserSetting from '@components/UserSetting'
-
-export default UserSetting
